requirements.txt
hydra-core==1.3.2
lightning==2.0.1.post0
numpy==1.24.2
tensorboard==2.12.2
torch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.1
wandb==0.15.0
